services/ws-hub/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
pydantic>=2.7.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# Observability (OTel)
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-exporter-otlp>=1.25.0
